Diagram update Jan 24, 2024
Minor changes to primary and foreign keys of each table. This was due to considerations regarding merging, cleaning, and data retention, erring on the side of retaining more data, not less. Also, much of the merge issue was regarding customer stats and customer demographics, and where to merge on. The index was used for the merge. 

Further normalization is likely needed for the credit card information, and further study, research, and consultation should be sought for how to comply with best data security practices.